requirements.txt
langchain>=0.1.0
langchain-community>=0.0.18
pypdf
pinecone
pyreadline3
openai
sentence-transformers
tqdm
pinecone-text
python-dotenv
scikit-learn
transformers
torch
streamlit
ollama